refactor(novu): extract searchNovuDocs helper for cleaner scaffold fixes NV-8302
Move docs search logic into tools/search-novu-docs.ts so support-agent.tsx only shows framework-specific tool wiring and approval config.
