+++ packages/novu/src/commands/init/templates/app-agent-langchain.template.spec.ts
@@ -17,6 +17,9 @@ describe('app-agent-langchain template', () => {
     expect(source).not.toMatch(/import\s*\{[^}]*\bagent\b[^}]*\}\s*from\s*'@novu\/framework'/);
     expect(source).not.toContain("from '@novu/framework/ai-sdk'");
     expect(source).toContain("import { tool } from '@langchain/core/tools'");
+    expect(source).toContain(
+      "import { searchNovuDocsIndex, searchNovuDocsInputSchema } from './tools/search-novu-docs'"
+    );
     expect(source).toContain('const searchNovuDocs = tool(');
     expect(source).toContain("toolCall.name === 'searchNovuDocs'");
     expect(activeImports).toMatch(/@langchain\/core/);
